+++ bitalarm-impl/src/test/java/com/dwarfeng/bitalarm/impl/configuration/ExceptionCodeOffsetConfiguration.java
@@ -15,11 +15,14 @@ public class ExceptionCodeOffsetConfiguration {
     private int subgradeExceptionCodeOffset;
     @Value("${bitalarm.exception_code_offset.snowflake}")
     private int snowflakeExceptionCodeOffset;
+    @Value("${bitalarm.exception_code_offset.dwarfeng_datamark}")
+    private int dwarfengDatamarkExceptionCodeOffset;
 
     @PostConstruct
     public void init() {
         ServiceExceptionCodes.setExceptionCodeOffset(exceptionCodeOffset);
         com.dwarfeng.subgrade.sdk.exception.ServiceExceptionCodes.setExceptionCodeOffset(subgradeExceptionCodeOffset);
         com.dwarfeng.sfds.sdk.util.ServiceExceptionCodes.setExceptionCodeOffset(snowflakeExceptionCodeOffset);
+        com.dwarfeng.datamark.util.ServiceExceptionCodes.setExceptionCodeOffset(dwarfengDatamarkExceptionCodeOffset);
     }
 }